Extract LazyRenderDelegate
Because "lazy" mode was the original mode, it is still the de facto default across the code base. We should try to be explicit when we're testing things in which mode, so this commit tries to extract and rename some objects, in particular the LazyRenderDelegate, to make it more explicit about what mode things are in.
