+++ packages/@glimmer/test-helpers/test/invocation-generation-test.ts
@@ -1,9 +1,9 @@
-import { TestEnvironmentRenderDelegate, RenderTest } from "@glimmer/test-helpers";
+import { LazyRenderDelegate, RenderTest } from "@glimmer/test-helpers";
 
 let renderTests: RenderTest;
 QUnit.module("Render Tests: buildComponent", {
   beforeEach() {
-    renderTests = new RenderTest(new TestEnvironmentRenderDelegate());
+    renderTests = new RenderTest(new LazyRenderDelegate());
   }
 });
